Reorganized some files, disabled Google button again.
