+++ gui/components/auth/oauth/oauthGoogle.tsx
@@ -1,9 +1,9 @@
 import * as React from "react";
-import OauthButton from "./forms/oauthButton";
+import OauthButton from "./oauthButton";
 
 export default function OauthGoogle() {
 
-    const localImagePath = "../../assets/images/google/btn_google_light_normal_ios.svg";
+    const localImagePath = "../../../assets/images/google/btn_google_light_normal_ios.svg";
     const provider = "google";
     const disabled = true;
 
@@ -14,5 +14,3 @@ export default function OauthGoogle() {
         </React.Fragment>
     )
 }
-
-
